+++ opentrons_sdk/helpers/helpers.py
@@ -1,4 +1,7 @@
 import json
+import math
+
+from opentrons_sdk.util.vector import Vector
 
 
 def unpack_coordinates(coordinates):
@@ -14,6 +17,39 @@ def flip_coordinates(coordinates, dimensions):
     x_size, y_size, z_size = unpack_coordinates(dimensions)
 
     return (x, y_size - y, z_size - z)
+
+
+def path_to_steps(p1, p2, increment=5, mode='absolute'):
+    """
+    given two points p1 and p2, this returns a list of
+    incremental positions or relative steps
+    """
+    Point = Vector
+    if not isinstance(p1, Vector):
+        Point = float
+
+    distance = Point(p2)
+    if mode is 'absolute':
+        distance = p2 - p1
+
+    if isinstance(p1, Vector):
+        travel = math.sqrt(
+            pow(distance[0], 2) + pow(distance[1], 2) + pow(distance[2], 2))
+    else:
+        travel = distance
+
+    divider = max(int(travel / increment), 1)
+    step = Point(distance / divider)
+
+    res = []
+    if mode is 'absolute':
+        for i in range(divider):
+            p1 = p1 + step
+            res.append(Point(p1))
+    else:
+        for i in range(divider):
+            res.append(step)
+    return res
 
 
 def import_calibration_json(json_string, robot, calibrated_top=False):
